requirement.txt
streamlit
tensorflow
numpy
pillow
scikit-image
matplotlib
